@@ -16,27 +16,24 @@ public class Command implements CommandExecutor {
     private final Plugin plugin = RPGDrop.getPlugin(RPGDrop.class);
 
     @Override
-    public boolean onCommand(@NotNull CommandSender sender, @NotNull org.bukkit.command. Command command, @NotNull String label, @NotNull String[] args) {
+    public boolean onCommand(@NotNull CommandSender sender, @NotNull org.bukkit.command.Command command, @NotNull String label, @NotNull String[] args) {
         if (args.length != 1) {
             sender.sendMessage(ChatColor.translateAlternateColorCodes('&', Objects.requireNonNull(plugin.getConfig().getString("illegalArgumentsMessage"))));
-        }
-        else if (args[0].equalsIgnoreCase("clear")) {
+        } else if (args[0].equalsIgnoreCase("clear")) {
             for (World world : Bukkit.getWorlds()) {
                 for (Item item : world.getEntitiesByClass(Item.class)) {
-                    if (!ItemOperations.isItemInProtectionList(item.getItemStack())) continue;
-                    ItemOperations.clearItem(item.getItemStack());
+                    if (!ItemOperations.isItemProtected(item.getItemStack())) continue;
+                    ItemOperations.clearProtection(item.getItemStack());
                 }
             }
             sender.sendMessage(ChatColor.translateAlternateColorCodes('&', Objects.requireNonNull(plugin.getConfig().getString("clearProtectionMessage"))));
-        }
-        else if (args[0].equalsIgnoreCase("reload")) {
+        } else if (args[0].equalsIgnoreCase("reload")) {
             plugin.reloadConfig();
             RPGDrop.config = plugin.getConfig();
             RPGDrop.configManager = new ConfigurationManager();
             RPGDrop.configManager.checkForDependencies();
             sender.sendMessage(ChatColor.translateAlternateColorCodes('&', Objects.requireNonNull(plugin.getConfig().getString("configReloadMessage"))));
-        }
-        else {
+        } else {
             sender.sendMessage(ChatColor.translateAlternateColorCodes('&', Objects.requireNonNull(plugin.getConfig().getString("illegalArgumentsMessage"))));
         }
         return true;
